services/lab-manager-v2/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-firestore==2.20.0
google-cloud-secret-manager==2.20.1
pydantic==2.10.3
bcrypt==4.1.2
pyjwt==2.8.0
email-validator==2.1.0
google-cloud-run==0.10.1
google-cloud-dns>=0.29.0
google-cloud-iam>=2.15.0
resend==2.0.0
